SSG-API-Testing-Application/Application Source Code/requirements.txt
Pillow~=8.2.0
requests~=2.25.1
cryptography~=3.4.7 